recruitment_assistant/requirements.txt
langgraph
langchain-openai
python-dotenv
fastapi
uvicorn[standard]
pytest
httpx
